+++ modules/Connection.js
@@ -1,6 +1,6 @@
 /* jshint -W058 */
 var d = require('describe-property');
-var isBinary = require('./utils/isBinary');
+var isBinary = require('bodec').isBinary;
 var decodeBase64 = require('./utils/decodeBase64');
 var encodeBase64 = require('./utils/encodeBase64');
 var stringifyQuery = require('./utils/stringifyQuery');